+++ src/Modules/Tours/Explorer.Tours.API/Public/Administration/ITourService.cs
@@ -9,18 +9,18 @@ namespace Explorer.Tours.API.Public.Administration;
 
 public interface ITourService {
     Result<List<TourDto>> GetByAuthorId(int id);
-    Result<TourDto> GetById(int id);
+    Result<TourDto> GetById(long id);
     Result<TourDto> Create(TourCreationDto tour);
     Result<TourDto> Update(TourDto tour, long id);
     Result Delete(int Id);
     Result UpdateTourEquipment(long tourId, List<long> equipmentId);
     Result<PagedResult<EquipmentDto>> GetTourEquipment(long tourId);
-    Result PublishTour(int tourId, double priceAmount, Currency currency);
+    Result PublishTour(long tourId, double priceAmount, Currency currency);
 
     Result<List<TourCardDto>> GetPublishedPagedTours(int page, int pageSize);
     Result<List<TourCardDto>> GetPublishedPagedToursFiltered(int page, int pageSize, double startLong, double endLong, double startLat, double endLat);
     Result<TourReviewDto> AddReview(TourReviewDto reviewDto);
     Result<TourTouristDto> GetForTouristById(long id, long touristId);
-    Result ArchiveTour(int tourId);
+    Result ArchiveTour(long tourId);
 
 }
